Client : Fix CollisionManager
Fix RectCollision

+++ PSK_DUNGREED/Client/Room_2.cpp
@@ -52,8 +52,8 @@ void CRoom_2::LateInit()
 	);
 
 	ScrollManager->SetMaxScroll(
-		TILECX *  (dynamic_cast<CTileMap*>(ObjectManager->GetObjectList(OBJ_TILEMAP)->front())->GetTileX() - 20)
-		, TILECY *  (dynamic_cast<CTileMap*>(ObjectManager->GetObjectList(OBJ_TILEMAP)->front())->GetTileY() - 13)
+		float(TILECX *  (dynamic_cast<CTileMap*>(ObjectManager->GetObjectList(OBJ_TILEMAP)->front())->GetTileX() - 20))
+		, float(TILECY *  (dynamic_cast<CTileMap*>(ObjectManager->GetObjectList(OBJ_TILEMAP)->front())->GetTileY() - 13))
 	);
 }
